Move enum to its own file
